refactor: Plugin helper functions to lib
